improve layout from buy confirmed component
